Update package-lock.json files in publish.sh
By default lerna doesn't seem to handle this - this runs npm install in
the packages which updates the package-lock.json files, then adds them
to the previous commit and pushed to github.

Not doing this caused a problem for me where if you run publish.sh twice
without something else happening to update the lockfiles, lerna fails to
publish because the npm install in publish.sh updates the lockfiles and
lerna won't run on a non-empty working tree.